requirements.txt
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
chardet==5.2.0
ftfy==6.2.0
python-dateutil==2.9.0
rapidfuzz==3.9.3
pydantic==2.7.1
pyfpgrowth==1.0
mlxtend==0.23.1
scikit-learn==1.4.2
networkx==3.3
joblib==1.4.2
scipy==1.13.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
Jinja2==3.1.4
openpyxl==3.1.2
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0
python-dotenv==1.0.1
